add agent picture

+++ src/main/java/com/tripthyst/webapp/service/AgentService.java
@@ -4,6 +4,6 @@ import com.tripthyst.webapp.model.AgentModel;
 
 public interface AgentService {
 
-    AgentModel getAgent(int id);
+    AgentModel getAgent(long id);
 
 }
